Add files via upload
pop up added to ask user if they are sure they want to delete contact

+++ usefulcode.js
@@ -120,7 +120,17 @@ function buildTable() {
 
     button.addEventListener("click", function()
     {
-      deleteContact(id);
+      var userPreference;
+
+      if (confirm("Do you want to delete contact?") == true) 
+      {
+         deleteContact(id);
+      } 
+      else 
+      {
+        userPreference = "Alright!";
+      }
+      //deleteContact(id);
       buildTable();
     });
 
@@ -179,7 +189,7 @@ function updateForm() {
 
       if(error == "")
       {
-        alert("contact successfully updated");
+        //alert("contact successfully updated");
         return false;
       }
 
@@ -199,7 +209,7 @@ function updateForm() {
 
 function deleteContact(record_id){
 
-  alert("Delete:" + record_id);
+  //alert("Delete:" + record_id);
   var deletePayload = '{"record_id" : "' + record_id + '", "user_id" : "' + window.sessionStorage.getItem("user_id") + '"}';
   var url = urlBegin + '/deleteContact' + urlEnding;
 
@@ -224,7 +234,7 @@ function deleteContact(record_id){
 
     if(error == "")
     {
-      alert("contact successfully deleted")
+      //alert("contact successfully deleted")
       //alert("userid: " + window.sessionStorage.getItem("user_id"));
       return false;
     }
